Fix to prevent rubber refs from being nuked
The figure can be adjusted when needed be but it's a nice fix for now. It would take a lot of ic investment to successfully knock out a large portion of rubber refs and that is if it goes uncontested. So static AA and fighters will still be a more effective measure and may even prevent them from getting blown up completely.